[3p sigin] Render a fallback signin dialog if popups are disabled (#7415)
Display a sign-in dialog as fallback for when we were blocked from
immediately calling window.open.

Safari has a much shorter navigator.userActivation.isActive timeout than
Chrome and Firefox, which means that too much time can elapse between
the guest asking for a sign-in and the host calling window.open.