[ve] Ensure output anchors open in new windows (#7421)
Since we are in an iframe with CSP we can't navigate to a different
origin. This breaks the case where Generate steps are asked to create
markdown links and output them as part of manual output. This updates
the Particle markdown renderer to ensure that there is a `_target` (as
well as `noopener` and `noreferrer` values) set.