[signin] Handle the /_app/ prefix in URL logic with explicit parameter instead of heuristic (#7423)
During some sign-ins from e.g. the signed-out opal page, we call
window.location.reload() because a fresh load is currently required for
signin.

However, sometimes our guest iframe URLs had been set to a URL that
lacked the shell guest /_app/ prefix, so window.location.reload() would
end up trying to reload with a shell host URL, and fail with a CSP
error.

That was happening because we were depending on a heuristic within
makeUrl to decide whether the guest prefix should be included, which was
whether it included a '?shared' parameter. Combining this with automatic
canonicalization logic that the router does meant that the iframe guest
sometimes got reset to a having a host url.

This PR replaces that heuristic with an explicit `guestPrefixed`
property that all calls to makeUrl must include. URLs being used for
within-guest navigations have it set to false, and ones for sharing etc.
have it set to true.