[lite] Fix broken lite mode snackbar (#7426)
The snackbar was not working in lite mode for a few reasons:

- We did not have snackbar event listeners on the prompt input element.
That's where prompt rejection errors come from.

- We were alternating between 3 different instances of the snackbar
element, depending on what lite mode ui state we were in. That meant
even if snackbar messages were passed to one snackbar, they could be
clobbered when that snackbar stopped being rendered. There was some
logic that was trying to graft messages between different snackbar
instances, but it didn't seem to be quite working. Anyway, it's much
simpler to just have one snackbar the whole time, which we now do.